+++ influxql/ast_test.go
@@ -601,7 +601,67 @@ func TestParseString(t *testing.T) {
 			stmt: `SELECT "cpu load" FROM "my\"series"`,
 		},
 		{
+			stmt: `SELECT "field with spaces" FROM "\"ugly\" db"."\"ugly\" rp"."\"ugly\" measurement"`,
+		},
+		{
 			stmt: `SELECT * FROM myseries`,
+		},
+		{
+			stmt: `DROP DATABASE "!"`,
+		},
+		{
+			stmt: `DROP RETENTION POLICY "my rp" ON "a database"`,
+		},
+		{
+			stmt: `CREATE RETENTION POLICY "my rp" ON "a database" DURATION 1d REPLICATION 1`,
+		},
+		{
+			stmt: `ALTER RETENTION POLICY "my rp" ON "a database" DEFAULT`,
+		},
+		{
+			stmt: `SHOW RETENTION POLICIES ON "a database"`,
+		},
+		{
+			stmt: `SHOW TAG VALUES WITH KEY IN ("a long name", short)`,
+		},
+		{
+			stmt: `DROP CONTINUOUS QUERY "my query" ON "my database"`,
+		},
+		{
+			stmt: `DELETE FROM "my db"."my rp"."my measurement"`,
+		},
+		{
+			stmt: `DROP SUBSCRIPTION "ugly \"subscription\" name" ON "\"my\" db"."\"my\" rp"`,
+		},
+		{
+			stmt: `CREATE SUBSCRIPTION "ugly \"subscription\" name" ON "\"my\" db"."\"my\" rp" DESTINATIONS ALL 'my host', 'my other host'`,
+		},
+		{
+			stmt: `SHOW MEASUREMENTS WITH MEASUREMENT =~ /foo/`,
+		},
+		{
+			stmt: `SHOW MEASUREMENTS WITH MEASUREMENT = "and/or"`,
+		},
+		{
+			stmt: `DROP USER "user with spaces"`,
+		},
+		{
+			stmt: `GRANT ALL PRIVILEGES ON "db with spaces" TO "user with spaces"`,
+		},
+		{
+			stmt: `GRANT ALL PRIVILEGES TO "user with spaces"`,
+		},
+		{
+			stmt: `SHOW GRANTS FOR "user with spaces"`,
+		},
+		{
+			stmt: `REVOKE ALL PRIVILEGES ON "db with spaces" FROM "user with spaces"`,
+		},
+		{
+			stmt: `REVOKE ALL PRIVILEGES FROM "user with spaces"`,
+		},
+		{
+			stmt: `CREATE DATABASE "db with spaces"`,
 		},
 	}
 
@@ -612,9 +672,13 @@ func TestParseString(t *testing.T) {
 			t.Fatalf("invalid statement: %q: %s", tt.stmt, err)
 		}
 
-		_, err = influxql.NewParser(strings.NewReader(stmt.String())).ParseStatement()
+		stmtCopy, err := influxql.NewParser(strings.NewReader(stmt.String())).ParseStatement()
 		if err != nil {
 			t.Fatalf("failed to parse string: %v\norig: %v\ngot: %v", err, tt.stmt, stmt.String())
+		}
+
+		if !reflect.DeepEqual(stmt, stmtCopy) {
+			t.Fatalf("statement changed after stringifying and re-parsing:\noriginal : %v\nre-parsed: %v\n", tt.stmt, stmtCopy.String())
 		}
 	}
 }